move trie tests to json_infra
